Rework Coinbase to use the new request controller
The websocket api of Coinbase is not reliably reporting pair prices,
therefore the rest api needs to be used. The rate limit of 10 requests
per second and the lack of an enpoint providing all prices at once,
make it necessary to query each configured pair prior to the GetPrices
call to be able to provide all requested prices in time.

The new version of the Coinbase provider uses the request controller
and constantly updates the configured prices in batches of 10 requests
at once with a short time between the batches to not exceed the rate
limits. The interval of the request lopp is set automatically according
to the amount of configured currency pairs.